Add time-shifting logic to make the "better" books be published newer than the "bad" ones.
Most feed readers will order from newest to oldest, so we'd like the better ranked books to be published sooner than the old ones. Since we parse in order of "best to worst", we need to offset our publish time so that we don't end up with the "worst" books as the newest articles in the feed.